Testing if die has all the default values
The test helps to check if all the default values are set properly and is accessible, the test checks die has a default name of d6, number of sides (max) is set to 6 and when the roll method is called via getcurrentside the number is between the range of 1 to 6. The test assists in identifying if we missed to give any value.