Test if die can set custom sides
The test helps to identify if we are able to set custom die sides, we are passing in the values in the test and setting it as max value of the die for which the random number between 1 to the max value passed will be generated and tested that the result is in the range of 1 to the number we set. We are also testing the name by appending the number of sides to d and comparing it to the string passed in the test.