Testing if the die roll method is between 1 to custom side we set
the test has a max value passed in which then rolls the die thousand time for each number and checks if the result is between 1 to the number of sides we passed. The test helps to easily check if there are any values returned by the roll that we haven't accounted for and the result is always in between the range.