Testing if Current side returns the value
Just like previous two tests after setting variables to private the test checks if they are accessible via getter and it returns a value. For the given test we are testing it with default value, the side is returned between default values (1 to 6) where 6 is the default max that we've set. The test helps to check if we are able to access the currenside up that is given by roll method