UI Tests
Added Few tests for UI, Testing if the label is displayed, where the test checks if text on label is readable that was written when programmed e.g. "test" will read the text and compare it to the given value in test.

The next test tests that radio buttons are displayed e.g. a radio button is named "radio1" it'll compare the text with the test text and see if it matches

The next test tests that radio button can be checked, with invoking a method called "isChecked" and passing in the value that we want (bool) if checked it'll be true if not the value will be false. where the test can conclude if the radio button can be checked or not.

The next test is about buttons if they are displayed properly, its a little tricky to check them but the test checks the text written on button and compares it with our text passed in the test e.g. A button containing name "Button 1", "Button 2", "Button 3" etc. and we pass in argument text contains Button, it will check for all the buttons that contains "Button * " this way we can make sure that all buttons are displayed and is accessible.

Two buttons methods are completed where  user can select any radio button where the value is passed into the die class and returns the current value, if roll once is clicked only one value is displayed, if roll two button is clicked two values are displayed on screen. this is done by making two dice objects.